Clean up.
