--- modules/substrate/fuzzer/Cargo.toml
@@ -1,61 +1,0 @@
-[package]
-name = "substrate-bridge-fuzzer"
-version = "0.1.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "1.3.1", default-features = false }
-finality-grandpa = { version = "0.12.3", default-features = false }
-hash-db = { version = "0.15.2", default-features = false }
-serde = { version = "1.0", optional = true }
-pallet-substrate-bridge = {path = ".." }
-
-# Bridge Dependencies
-
-bp-header-chain = { path = "../../../primitives/header-chain", default-features = false }
-bp-runtime = { path = "../../../primitives/runtime", default-features = false }
-
-# Substrate Dependencies
-
-frame-support = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-frame-system = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-sp-finality-grandpa = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-sp-std = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-sp-trie = { git = "https://github.com/paritytech/substrate.git", branch = "master" , default-features = false }
-
-bp-test-utils = {path = "../../../primitives/test-utils" }
-sp-core = { git = "https://github.com/paritytech/substrate.git", branch = "master" }
-sp-io = { git = "https://github.com/paritytech/substrate.git", branch = "master" }
-sp-state-machine = { git = "https://github.com/paritytech/substrate.git", branch = "master" }
-
-honggfuzz = ""
-
-[features]
-default = ["std"]
-std = [
-	"bp-header-chain/std",
-	"bp-runtime/std",
-	"bp-header-chain/std",
-	"codec/std",
-	"finality-grandpa/std",
-	"frame-support/std",
-	"frame-system/std",
-	"hash-db/std",
-	"serde",
-	"sp-finality-grandpa/std",
-	"sp-runtime/std",
-	"sp-std/std",
-	"sp-trie/std",
-]
-runtime-benchmarks = []
-
-[[bin]]
-name = "storage_proof_fuzzer"
-path = "src/storage_proof_fuzzer.rs"
-
-[workspace]
